remove built-in clock, add clock plugin.
git-svn-id: https://g15daemon.svn.sourceforge.net/svnroot/g15daemon/trunk@109 ab83d0d9-5618-0410-a047-8efd564e0644

+++ g15daemon-wip/g15daemon/g15daemon.h
@@ -183,7 +183,7 @@ struct lcdlist_s
     lcdnode_t *current;
     void *(*keyboard_handler)(void*);
     struct passwd *nobody;
-    int numclients;
+    unsigned long numclients;
 }lcdlist_s;
 
 pthread_mutex_t lcdlist_mutex;
